Renamed module from "converter" to "edit-text-utils" and incapsulated Converter classes and Verifier classes to corresponding packages
