--- converter/src/main/java/ru/touchin/converter/verifiers/PointVerifier.kt
@@ -1,25 +1,0 @@
-package ru.touchin.converter.verifiers
-
-import android.util.Log
-import ru.touchin.converter.commands.Command
-import java.math.BigDecimal
-
-/**
- * Verifies separator existence in text input.
- * @param separatorChar character to check.
- */
-class PointVerifier(separatorChar: Char = '.') : Verifier<BigDecimal> { // todo fix regular expression to catch repeating zeros
-
-    private val pattern = "[^$separatorChar]*[$separatorChar]?[^$separatorChar]*".toRegex()
-
-    override fun verify(text: String): Command<BigDecimal> {
-        Log.d("verify", "text:${text}")
-        return if (pattern.matches(text) == true) {
-            Command.Success()
-        } else {
-
-            Command.Fallback()
-        }
-    }
-
-}
