+++ edit-text-utils/src/main/java/ru/touchin/edittextutils/verifier/verifiers/Verifier.kt
@@ -1,0 +1,13 @@
+package ru.touchin.edittextutils.verifier.verifiers
+
+import ru.touchin.edittextutils.commands.Command
+import java.math.BigDecimal
+
+/**
+ * Basic verifier class used as step of input verification
+ */
+interface Verifier<T : BigDecimal> {
+
+    fun verify(text: String): Command<T>
+
+}
