+++ modules.gradle
@@ -5,7 +5,7 @@ if (gradle.ext.has('componentsRoot')) {
     rootDir = settingsDir
 }
 
-include ':converter'
+include ':edit-text-utils'
 include ':logging'
 include ':utils'
 include ':navigation'
@@ -18,7 +18,7 @@ include ':recyclerview-adapters'
 include ':kotlin-extensions'
 include ':templates'
 
-project(':converter').projectDir = new File(rootDir, 'converter')
+project(':edit-text-utils').projectDir = new File(rootDir, 'edit-text-utils')
 project(':utils').projectDir = new File(rootDir, 'utils')
 project(':logging').projectDir = new File(rootDir, 'logging')
 project(':navigation').projectDir = new File(rootDir, 'navigation')
